Support Windows System Linux WSL
The script gpg-agent-relay.sh needs to have
gpg --card-status done on a windows termainal
before it works and then run a "Ubuntu" terminal.
Also, for some strange reason `rm .gnupg/S.gpg-agent*`
fails sometimes with "no such files" even when they
are present. This is why the `ls -l .gnupg` output
is still in the script.