Add Host for devices on LAN for sshing into devices.
Generally Host's have two names the device name such as 3900x and the
"ip" name. Which for 3900x is ip101. The 101 is the last decimal byte of
it's local address, which for 3900x is 192.168.1.101.

In addition another pair of names with the pubic hostname, 23.119.164.150,
is also available. So 3900x can be ssh'd into from outside the network with
3900z or ip101z.

Examples:

  ssh wink@3900x
  ssh wink@3900xz

or

  ssh wink@ip101
  ssh wink@ip101z