feat: Add updating .tar.gz or .tar files
Added function get_file_and_extract which taks a file_name as the first
parameter and an extension as the second parameter. It assumes between
the file_name and the extension is "-x86_64_unknown-linux". The function
is called first assuming ".tar.gz" extension as that has been the default.
If that fails it tries ".tar" as the extension.

Also removed the "-e" from the set at the beginning so exiting doesn't
happen on failures, but I do add the "set -e" when doing the finishing
work of changing the mode to executable, remove old ~/bin/lighthouse
symlink and then create a new symlink.